+ fix vendor code to long

+++ src/main/java/ru/demo/marketplaceservice/dto/ProductDto.java
@@ -18,6 +18,6 @@ public class ProductDto {
     private String name;
     private Double price;
     private String currency;
-    private Integer vendorCode;
+    private Long vendorCode;
     private Boolean deleted;
 }
